2+4s
movement
and
movementBLUE
are the 2+4 files, red is functional, blue has had minimal testing but there is potential

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/SourceCode/Auton/BLUEbackdropCycle.java
@@ -43,7 +43,7 @@ public class BLUEbackdropCycle extends LinearOpMode {
     public DcMotorEx rightSlide;
 
     public HuskyStackDetection detect;
-    int dist = 0;
+    double dist = 0;
     private HuskyLens huskyLens;
 
     public OpenCvCamera webcam;
